Camera: Restore scene mode to default
When user chooses to restore the default settings, the scene mode is
not being reset to the default value.
The mismatch in the string value of default scene mode in camera
parameters and in camera preferences causing this issue.
So, match the default scene mode string value in preferences to the
camera parameters string value.

Change-Id: I58a74b218e1bc09d8a4c56bdc895a2b4982eff4f
CRs-fixed: 353725